refactor : 노선 도메인이 구간 일급컬렉션을 가지도록 리팩토링
1. 도메인 수정 및 ID가 없을 때 임시로 도메인 객체를 생성할 수 있도록 EmptySections null객체 생성
2. 테스트 수정
3. LineService가 컨트롤러로 Dto를 넘기도록 수정
4. LineService가 SectionService를 포함하도록 수정

+++ src/main/java/wooteco/subway/controller/web/line/LineRequest.java
@@ -14,12 +14,12 @@ public class LineRequest {
     @NotNull
     private Long downStationId;
     @Positive
-    private int distance;
+    private Long distance;
 
     public LineRequest() {
     }
 
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -43,7 +43,7 @@ public class LineRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
