+++ src/main/java/wooteco/subway/controller/web/section/SectionRequest.java
@@ -8,12 +8,12 @@ public class SectionRequest {
     @NotNull
     private Long downStationId;
     @NotNull
-    private int distance;
+    private Long distance;
 
     public SectionRequest() {
     }
 
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+    public SectionRequest(Long upStationId, Long downStationId, Long distance) {
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
@@ -27,7 +27,7 @@ public class SectionRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
